Make DotnetOperations a partial class
- This splits out the majority of methods in the static class into their own files.
